refactor(server): manifest type
