+++ packages/twenty-server/src/engine/metadata-modules/flat-entity/types/add-suffix-to-entity-many-to-one-properties.type.ts
@@ -1,7 +1,7 @@
 import { type AllMetadataName } from 'twenty-shared/metadata';
 
 import { type MetadataManyToOneJoinColumn } from 'src/engine/metadata-modules/flat-entity/types/metadata-many-to-one-join-column.type';
-import { type RemoveSuffix } from 'src/engine/workspace-manager/workspace-migration/workspace-migration-builder/types/remove-suffix.type';
+import { RemoveSuffix } from 'twenty-shared/types';
 
 export type AddSuffixToEntityManyToOneProperties<
   TEntity,
