+++ packages/twenty-server/src/engine/metadata-modules/flat-entity/types/add-suffix-to-entity-one-to-many-properties.type.ts
@@ -1,6 +1,6 @@
 import { type ExtractEntityOneToManyEntityRelationProperties } from 'src/engine/metadata-modules/flat-entity/types/extract-entity-one-to-many-entity-relation-properties.type';
 import { type SyncableEntity } from 'src/engine/workspace-manager/types/syncable-entity.interface';
-import { type RemoveSuffix } from 'src/engine/workspace-manager/workspace-migration/workspace-migration-builder/types/remove-suffix.type';
+import { RemoveSuffix } from 'twenty-shared/types';
 
 export type AddSuffixToEntityOneToManyProperties<
   TEntity,
