+++ packages/twenty-shared/src/types/RemoveSuffix.type.ts
